fix: mdsvex compiler
