+++ package.json
@@ -28,10 +28,10 @@
 		"@iconify/svelte": "3.1.3",
 		"@iconify/types": "2.0.0",
 		"@macfja/svelte-persistent-store": "2.2.1",
-		"@playwright/test": "1.32.3",
+		"@playwright/test": "1.33.0",
 		"@svelte-put/clickoutside": "2.0.0",
 		"@svelte-put/intersect": "2.0.0",
-		"@sveltejs/adapter-vercel": "2.4.2",
+		"@sveltejs/adapter-vercel": "2.4.3",
 		"@sveltejs/kit": "1.15.9",
 		"@tailwindcss/typography": "0.5.9",
 		"@threlte/core": "5.0.9",
@@ -81,7 +81,7 @@
 		"swiper": "9.2.4",
 		"tailwind-scrollbar": "3.0.0",
 		"tailwindcss": "3.3.2",
-		"three": "0.151.3",
+		"three": "0.152.0",
 		"tslib": "2.5.0",
 		"typescript": "5.0.4",
 		"vite": "4.3.3",
