+++ src/lib/utils/index.ts
@@ -2,7 +2,7 @@ export * from './ascii';
 export * from './date';
 export * from './formatters';
 export * from './host';
-export * from './mdsvex-compiler';
+export * from './mdsvex';
 export * from './reading-time';
 export * from './reduced-motion';
 export * from './routing';
