+++ src/routes/blog/+page.server.ts
@@ -1,5 +1,7 @@
 import type { ICMSData, IPost } from '$lib/ts';
 
+export const prerender = false;
+
 export async function load({ fetch }) {
 	const [{ posts }, { posts: featured }] = await Promise.all([
 		fetch('/api/posts').then<{ posts: ICMSData<IPost[]> }>(res => res.json()),
